requirements.txt
# Ketter 3.0 - Python Dependencies
# Install with: pip install -r requirements.txt

# ============================================
# Web Framework (FastAPI)
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# Database (PostgreSQL)
# ============================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ============================================
# Job Queue (Redis + RQ)
# ============================================
redis==5.0.1
rq==1.16.1

# ============================================
# File Operations & Checksums
# ============================================
# hashlib is built-in, but we need these for large files
aiofiles==23.2.1
tqdm==4.66.1

# ============================================
# PDF Report Generation
# ============================================
reportlab==4.0.9

# ============================================
# Testing
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For FastAPI testing
requests==2.31.0  # For test_large_files.py

# ============================================
# Development Tools
# ============================================
python-dotenv==1.0.1
python-multipart==0.0.6  # For file uploads
pyyaml==6.0.1  # For ketter.config.yml parsing

# ============================================
# Logging & Monitoring
# ============================================
python-json-logger==2.0.7

# ============================================
# Type Hints
# ============================================
types-redis==4.6.0.20240106
